add: db_init loading of companies data.
- load companies into state in db_init()
- fix wrong integer sizes in StockVal & Company

+++ src/libtrader/ds/generic/company.rs
@@ -2,7 +2,7 @@ use postgres_types::{ToSql, FromSql};
 
 #[derive(Default, PartialEq, Debug, ToSql, FromSql)]
 pub struct Company {
-    pub id: i8,
+    pub id: i64,
     pub symbol: String,
     pub isin: String,
     pub company_name: String,
@@ -10,5 +10,5 @@ pub struct Company {
     pub sector: String,
     pub industry: String,
     pub primary_sic_code: String,
-    pub employees: i8,
+    pub employees: i64,
 }
