+++ src/libtrader/ds/generic/stock_val.rs
@@ -2,10 +2,10 @@ use postgres_types::{ToSql, FromSql};
 
 #[derive(Default, PartialEq, Debug, ToSql, FromSql)]
 pub struct StockVal {
-    pub id: i8,
+    pub id: i64,
     pub isin: String,
-    pub time_since_epoch: i32,
-    pub ask_price: i8,
-    pub bid_price: i8,
-    pub volume: i8
+    pub time_since_epoch: i64,
+    pub ask_price: i64,
+    pub bid_price: i64,
+    pub volume: i64
 }
